Improved slideshows
You can no longer spam click the slideshows.

It's hard to reproduce but I swear that if you click the button right as the image transition finishes, the click won't register. Using a print statement, I saw that the element wasn't running the changeSlide function at all, so I'm thinking this might be an issue with chrome itself.